Resolve circular dependency in Oncoprint
