+++ src/pages/staticPages/tools/oncoprinter/OncoprinterUtils.ts
@@ -2,8 +2,8 @@ import {observable} from "mobx";
 import AppConfig from "appConfig";
 import {default as OncoprinterStore} from "./OncoprinterStore";
 import _ from "lodash";
-import {GeneticTrackDatum, GeneticTrackDatum_Data,} from "../../../../shared/components/oncoprint/Oncoprint";
-import {percentAltered} from "../../../../shared/components/oncoprint/OncoprintUtils";
+import {GeneticTrackDatum_Data,} from "../../../../shared/components/oncoprint/Oncoprint";
+import {GeneticTrackDatum, percentAltered} from "../../../../shared/components/oncoprint/OncoprintUtils";
 import {AlterationTypeConstants} from "../../../../shared/lib/StoreUtils";
 import {cna_profile_data_to_string} from "../../../../shared/lib/oql/AccessorsForOqlFilter";
 import {fillGeneticTrackDatum, OncoprintMutationType} from "../../../../shared/components/oncoprint/DataUtils";
