+++ src/shared/components/oncoprint/Oncoprint.tsx
@@ -1,63 +1,21 @@
 import * as React from "react";
-import OncoprintJS, {TrackId, CustomTrackOption} from "oncoprintjs";
-import {GenePanelData, MolecularProfile} from "../../api/generated/CBioPortalAPI";
+import OncoprintJS, {TrackId} from "oncoprintjs";
+import {MolecularProfile} from "../../api/generated/CBioPortalAPI";
 import {observer} from "mobx-react";
 import {computed} from "mobx";
 import {transition} from "./DeltaUtils";
 import _ from "lodash";
-import {
-    ExtendedAlteration
-} from "../../../pages/resultsView/ResultsViewPageStore";
+import {ExtendedAlteration} from "../../../pages/resultsView/ResultsViewPageStore";
 import "./styles.scss";
 import {AnnotatedMutation, AnnotatedNumericGeneMolecularData} from "../../lib/oql/AccessorsForOqlFilter";
-
-export type ClinicalTrackDatum = {
-    attr_id: string;
-    study_id: string;
-    sample?:string;
-    patient?:string;
-    uid: string;
-    attr_val_counts: {[val:string]:number};
-    attr_val?: string|number|ClinicalTrackDatum["attr_val_counts"];
-    na?:boolean;
-};
-
-export type ClinicalTrackSpec = {
-    key: string; // for efficient diffing, just like in React. must be unique
-    label: string;
-    description: string;
-    data: ClinicalTrackDatum[];
-    altered_uids?:string[];
-    na_legend_label?:string;
-    na_tooltip_value?:string; // If given, then show a tooltip over NA columns that has this value
-    custom_options?:CustomTrackOption[];
-} & ({
-    datatype: "counts";
-    countsCategoryLabels:string[];
-    countsCategoryFills:string[];
-} | {
-    datatype: "number";
-    numberRange:[number, number];
-    numberLogScale:boolean;
-} | {
-    datatype: "string";
-    category_to_color?:{[category:string]:string}
-});
-
-export interface IBaseHeatmapTrackDatum {
-    profile_data: number|null;
-    sample?: string;
-    patient: string;
-    study_id: string;
-    uid: string;
-    na?:boolean;
-}
-export interface IGeneHeatmapTrackDatum extends IBaseHeatmapTrackDatum {
-    hugo_gene_symbol: string;
-}
-export interface IGenesetHeatmapTrackDatum extends IBaseHeatmapTrackDatum {
-    geneset_id: string;
-}
+import {
+    CLINICAL_TRACK_GROUP_INDEX,
+    ClinicalTrackSpec,
+    GENETIC_TRACK_GROUP_INDEX,
+    GeneticTrackSpec,
+    IGeneHeatmapTrackSpec,
+    IGenesetHeatmapTrackSpec
+} from "./OncoprintUtils";
 
 export type GeneticTrackDatum_Data =
     Pick<ExtendedAlteration&AnnotatedMutation&AnnotatedNumericGeneMolecularData,
@@ -66,63 +24,6 @@ export type GeneticTrackDatum_Data =
         "alterationSubType" | "value" | "mutationType" | "isHotspot" | "entrezGeneId" | "putativeDriver" | "mutationStatus">;
 
 export type GeneticTrackDatum_ProfiledIn = {genePanelId?:string, molecularProfileId:string};
-
-export type GeneticTrackDatum = {
-    trackLabel: string;
-    sample?:string;
-    patient:string;
-    study_id:string;
-    uid:string;
-    data:GeneticTrackDatum_Data[];
-    profiled_in?: GeneticTrackDatum_ProfiledIn[];
-    not_profiled_in?:GeneticTrackDatum_ProfiledIn[];
-    na?: boolean;
-    disp_mut?:string;
-    disp_cna?:string;
-    disp_mrna?:string;
-    disp_prot?:string;
-    disp_fusion?:boolean;
-    disp_germ?:boolean;
-};
-
-export type GeneticTrackSpec = {
-    key: string; // for efficient diffing, just like in React. must be unique
-    label: string;
-    sublabel?: string;
-    oql?: string; // OQL corresponding to the track
-    info: string;
-    infoTooltip?:string;
-    data: GeneticTrackDatum[];
-    expansionCallback?: () => void;
-    removeCallback?: () => void;
-    expansionTrackList?: GeneticTrackSpec[];
-    labelColor?: string;
-};
-
-interface IBaseHeatmapTrackSpec {
-    key: string; // for efficient diffing, just like in React. must be unique
-    label: string;
-    molecularProfileId: string; // source
-    molecularAlterationType: MolecularProfile["molecularAlterationType"];
-    datatype: MolecularProfile["datatype"];
-    data: IBaseHeatmapTrackDatum[];
-    trackGroupIndex: number;
-}
-export interface IGeneHeatmapTrackSpec extends IBaseHeatmapTrackSpec {
-    data: IGeneHeatmapTrackDatum[];
-    onRemove: () => void;
-    info?: string;
-    labelColor?: string;
-}
-export interface IGenesetHeatmapTrackSpec extends IBaseHeatmapTrackSpec {
-    data: IGenesetHeatmapTrackDatum[];
-    trackLinkUrl: string | undefined;
-    expansionTrackList: IGeneHeatmapTrackSpec[];
-    expansionCallback: () => void;
-}
-
-export const GENETIC_TRACK_GROUP_INDEX = 1;
-export const CLINICAL_TRACK_GROUP_INDEX = 0;
 
 export interface IOncoprintProps {
     oncoprintRef?:(oncoprint:OncoprintJS<any>)=>void;
